Ball movement enhancements
Ball no longer has magical velocity which increases and decreases on the x based on button input.
Ball now has rotational speed which on collision with floors calculates the increase to speed.
Allows ball to spin in the air and land with a boost.
Ball now also moves in relation to floor angle instead of only on the x.